requirements.txt
discord.py
openai
python-dotenv
requests